+++ packages/extension-base/src/types/metadata.ts
@@ -1,0 +1,10 @@
+// Copyright 2019-2022 @subwallet/extension-base
+// SPDX-License-Identifier: Apache-2.0
+
+export interface ResponseShortenMetadata {
+  txMetadata: string;
+}
+
+export interface ResponseMetadataHash {
+  metadataHash: string;
+}