+++ packages/extension-web-ui/src/hooks/ledger/useLedger.ts
@@ -1,13 +1,11 @@
-// Copyright 2019-2022 @subwallet/extension-koni-ui authors & contributors
+// Copyright 2019-2022 @subwallet/extension-web-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { _ChainInfo } from '@subwallet/chain-list/types';
 import { LedgerNetwork } from '@subwallet/extension-base/background/KoniTypes';
 import { _isChainEvmCompatible } from '@subwallet/extension-base/services/chain-service/utils';
-import { EVMLedger, SubstrateLedger } from '@subwallet/extension-web-ui/connector';
-import { isLedgerCapable } from '@subwallet/extension-web-ui/constants';
+import { EVMLedger, SubstrateGenericLedger, SubstrateLegacyLedger, SubstrateMigrationLedger } from '@subwallet/extension-web-ui/connector';
+import { isLedgerCapable, ledgerIncompatible, NotNeedMigrationGens } from '@subwallet/extension-web-ui/constants';
 import { useSelector } from '@subwallet/extension-web-ui/hooks';
-import useGetSupportedLedger from '@subwallet/extension-web-ui/hooks/ledger/useGetSupportedLedger';
 import { Ledger } from '@subwallet/extension-web-ui/types';
 import { convertLedgerError } from '@subwallet/extension-web-ui/utils';
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
@@ -16,6 +14,7 @@ import { AccountOptions, LedgerAddress, LedgerSignature } from '@polkadot/hw-led
 import { assert } from '@polkadot/util';
 
 import useTranslation from '../common/useTranslation';
+import useGetSupportedLedger from './useGetSupportedLedger';
 
 interface StateBase {
   isLedgerCapable: boolean;
@@ -46,23 +45,35 @@ const getNetwork = (ledgerChains: LedgerNetwork[], slug: string, isEthereumNetwo
   return ledgerChains.find((network) => network.slug === slug || (network.isEthereum && isEthereumNetwork));
 };
 
-const retrieveLedger = (slug: string, ledgerChains: LedgerNetwork[], isEthereumNetwork: boolean): Ledger => {
+const retrieveLedger = (slug: string, ledgerChains: LedgerNetwork[], isEthereumNetwork: boolean, forceMigration: boolean): Ledger => {
   const { isLedgerCapable } = baseState;
 
-  assert(isLedgerCapable, 'Incompatible browser, only Chrome is supported');
+  assert(isLedgerCapable, ledgerIncompatible);
 
   const def = getNetwork(ledgerChains, slug, isEthereumNetwork);
 
   assert(def, 'There is no known Ledger app available for this chain');
 
-  if (def.isEthereum) {
-    return new EVMLedger('webusb');
+  if (def.isGeneric) {
+    if (def.isEthereum) {
+      return new EVMLedger('webusb', def.slip44);
+    } else {
+      return new SubstrateGenericLedger('webusb', def.slip44);
+    }
   } else {
-    return new SubstrateLedger('webusb', def.network);
+    if (!forceMigration) {
+      return new SubstrateLegacyLedger('webusb', def.network);
+    } else {
+      if (NotNeedMigrationGens.includes(def.genesisHash)) {
+        return new SubstrateGenericLedger('webusb', def.slip44);
+      } else {
+        return new SubstrateMigrationLedger('webusb', def.slip44);
+      }
+    }
   }
 };
 
-export function useLedger (slug?: string, active = true): Result {
+export function useLedger (slug?: string, active = true, isSigning = false, forceMigration = false): Result {
   const { t } = useTranslation();
 
   const ledgerChains = useGetSupportedLedger();
@@ -75,6 +86,10 @@ export function useLedger (slug?: string, active = true): Result {
 
     const chainInfo = chainInfoMap[slug];
 
+    if (!chainInfo) {
+      return false;
+    }
+
     return _isChainEvmCompatible(chainInfo);
   }, [chainInfoMap, slug]);
 
@@ -86,6 +101,7 @@ export function useLedger (slug?: string, active = true): Result {
   const [refreshLock, setRefreshLock] = useState(false);
   const [warning, setWarning] = useState<string | null>(null);
   const [error, setError] = useState<string | null>(null);
+
   const [ledger, setLedger] = useState<Ledger| null>(null);
 
   const getLedger = useCallback(() => {
@@ -104,14 +120,14 @@ export function useLedger (slug?: string, active = true): Result {
       }
 
       try {
-        return retrieveLedger(slug, ledgerChains, isEvmNetwork);
+        return retrieveLedger(slug, ledgerChains, isEvmNetwork, forceMigration);
       } catch (error) {
         setError((error as Error).message);
       }
     }
 
     return null;
-  }, [refreshLock, slug, active, ledgerChains, isEvmNetwork]);
+  }, [refreshLock, slug, active, ledgerChains, isEvmNetwork, forceMigration]);
 
   const appName = useMemo(() => {
     const unknownNetwork = 'unknown network';
@@ -121,18 +137,26 @@ export function useLedger (slug?: string, active = true): Result {
     }
 
     const chainInfo = chainInfoMap[slug];
-    const isEthereumNetwork = _isChainEvmCompatible(chainInfo);
-    const { appName } = getNetwork(ledgerChains, slug, isEthereumNetwork) || { appName: unknownNetwork };
+    const isEthereumNetwork = chainInfo ? _isChainEvmCompatible(chainInfo) : false;
+    const { appName, isEthereum, isGeneric } = getNetwork(ledgerChains, slug, isEthereumNetwork) || { appName: unknownNetwork, isGeneric: true };
+
+    if (!isGeneric && forceMigration && !isEthereum) {
+      if (NotNeedMigrationGens.includes(chainInfo?.substrateInfo?.genesisHash || '')) {
+        return ledgerChains[0].appName;
+      } else {
+        return 'Polkadot Migration';
+      }
+    }
 
     return appName;
-  }, [chainInfoMap, ledgerChains, slug]);
+  }, [chainInfoMap, forceMigration, ledgerChains, slug]);
 
   const refresh = useCallback(() => {
     setRefreshLock(true);
   }, []);
 
   const handleError = useCallback((error: Error, expandError = true) => {
-    const convertedError = convertLedgerError(error, t, appName, expandError);
+    const convertedError = convertLedgerError(error, t, appName, isSigning, expandError);
     const message = convertedError.message;
 
     switch (convertedError.status) {
@@ -148,7 +172,7 @@ export function useLedger (slug?: string, active = true): Result {
         setWarning(null);
         setError(null);
     }
-  }, [appName, t]);
+  }, [appName, isSigning, t]);
 
   const getAllAddress = useCallback(async (start: number, end: number): Promise<LedgerAddress[]> => {
     const ledger_ = getLedger();
@@ -197,12 +221,12 @@ export function useLedger (slug?: string, active = true): Result {
       });
   }, [getLedger, ledger, t]);
 
-  const signTransaction = useCallback(async (message: Uint8Array, accountOffset?: number, addressOffset?: number, accountOption?: Partial<AccountOptions>): Promise<LedgerSignature> => {
+  const signTransaction = useCallback(async (message: Uint8Array, metadata: Uint8Array, accountOffset?: number, addressOffset?: number, accountOption?: Partial<AccountOptions>): Promise<LedgerSignature> => {
     if (ledger) {
       return new Promise((resolve, reject) => {
         setError(null);
 
-        ledger.signTransaction(message, accountOffset, addressOffset, accountOption)
+        ledger.signTransaction(message, metadata, accountOffset, addressOffset, accountOption)
           .then((result) => {
             resolve(result);
           })
@@ -268,7 +292,7 @@ export function useLedger (slug?: string, active = true): Result {
           console.error(error);
         });
     }, 300);
-  }, [slug, ledgerChains, t, active, chainInfoMap, appName, handleError, getLedger]);
+  }, [slug, t, active, handleError, getLedger]);
 
   useEffect(() => {
     destroyRef.current = () => {